requirements.txt
appdirs==1.4.0
Django==1.10.5
django-tinymce==2.6.0
django-webpack-loader==0.4.1
djangorestframework==3.5.3
MySQL-python==1.2.5
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
uWSGI==2.0.14
